[powerpc] Update parameters for csum_tcpudp_magic & csum_tcpudp_nofold
Message-id: <1490816019-8911-14-git-send-email-gduarte@redhat.com>
Patchwork-id: 175777
O-Subject: [RHEL7.4 PATCH BZ 1430834 13/16] arch/powerpc: Update parameters for csum_tcpudp_magic & csum_tcpudp_nofold
Bugzilla: 1430834
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ 1430834
Upstream Status: f9d4286b9516b02e795214412d36885f572b57ad

commit f9d4286b9516b02e795214412d36885f572b57ad
Author: Ivan Vecera <ivecera@redhat.com>
Date:   Thu Oct 27 16:30:06 2016 +0200

    arch/powerpc: Update parameters for csum_tcpudp_magic & csum_tcpudp_nofold

    Commit 01cfbad "ipv4: Update parameters for csum_tcpudp_magic to their
    original types" changed parameters for csum_tcpudp_magic and
    csum_tcpudp_nofold for many platforms but not for PowerPC.

    Fixes: 01cfbad "ipv4: Update parameters for csum_tcpudp_magic to their original types"
    Cc: Alexander Duyck <aduyck@mirantis.com>
    Acked-by: Alexander Duyck <alexander.h.duyck@intel.com>

+++ arch/powerpc/include/asm/checksum.h
@@ -76,10 +76,8 @@ static inline __sum16 ip_compute_csum(const void *buff, int len)
 	return csum_fold(csum_partial(buff, len, 0));
 }
 
-static inline __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
-                                     unsigned short len,
-                                     unsigned short proto,
-                                     __wsum sum)
+static inline __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr, __u32 len,
+					__u8 proto, __wsum sum)
 {
 #ifdef __powerpc64__
 	unsigned long s = (__force u32)sum;
@@ -106,10 +104,8 @@ static inline __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
  * computes the checksum of the TCP/UDP pseudo-header
  * returns a 16-bit checksum, already complemented
  */
-static inline __sum16 csum_tcpudp_magic(__be32 saddr, __be32 daddr,
-					unsigned short len,
-					unsigned short proto,
-					__wsum sum)
+static inline __sum16 csum_tcpudp_magic(__be32 saddr, __be32 daddr, __u32 len,
+					__u8 proto, __wsum sum)
 {
 	return csum_fold(csum_tcpudp_nofold(saddr, daddr, len, proto, sum));
 }
